feat: render the signup and signin pages with routes
