+++ src/App.tsx
@@ -1,9 +1,11 @@
-import { SignIn } from "./pages";
+import { Routes } from "./routes";
 
 function App() {
-  return <div>
-    <SignIn/>
-  </div>;
+  return (
+    <div>
+      <Routes />
+    </div>
+  );
 }
 
 export default App;
