+++ src/routes/index.tsx
@@ -1,0 +1,12 @@
+import { Switch, Route } from "react-router-dom";
+import { SignIn } from "../pages/SignIn";
+import { SignUp } from "../pages/SignUp";
+
+export const Routes = () => {
+  return (
+    <Switch>
+      <Route exact path="/" component={SignIn} />
+      <Route path="/signup" component={SignUp} />
+    </Switch>
+  );
+};
